backend/requirements.txt
# ============================================================================
# Plant Intel Backend Dependencies
# ============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase==2.9.0
# psycopg2-binary==2.9.9  # Not needed - using Supabase client instead
# sqlalchemy==2.0.23  # Not needed - using Supabase client instead

# Authentication
pyjwt==2.8.0
cryptography==42.0.8

# Task Queue
celery==5.3.4
redis==5.0.1
pyopenssl==24.0.0   # Forces compatibility with cryptography 45.x


# AI/ML
openai==1.3.7
anthropic==0.7.8
numpy>=1.26.4  # Python 3.13 compatible
pandas>=2.2.0  # Python 3.13 compatible
scikit-learn>=1.4.0  # Python 3.13 compatible

# Payment Processing
stripe==7.8.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.10.0  # Python 3.13 compatible
pydantic-settings>=2.6.0  # Python 3.13 compatible
python-dateutil==2.8.2
pytz==2023.3

# HTTP Client
httpx==0.27.0
requests==2.31.0

# Validation & Serialization
email-validator==2.1.0
python-jose==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.27.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0

# Security
python-decouple==3.8
bcrypt==4.1.2
